remove dataset-loader-manager
